Created a scoring system
Had issues with the scoring system: drew a line btwn the pipes (unnecessarily difficult without causing random glitches) and created a function where it would return either 1 or 0 if it passes thru the line. However, once it all adds up, it sums to 25; so I divide it by 25 at the end to get the actual score. Still trying to figure out how to reset the score without causing numerous bugs.